Moved the ebpf filter to a seperate file
